Prepared Pollymer for npm and jspm, bumped version to 1.1.3
