+++ .npmignore
@@ -1,0 +1,3 @@
+node_modules/
+jspm_packages/
+src/
