+++ CHANGELOG.md
@@ -5,3 +5,4 @@ v. 1.0.0 (01-31-2013) - Initial Release.
 v. 1.1.0 (04-28-2014) - Support for RequireJS (AMD).
 v. 1.1.1 (05-10-2014) - Fixed crashing issue when abort() called on fresh object.
 v. 1.1.2 (05-10-2014) - Adding access to "lastRequest" info.
+v. 1.1.3 (11-27-2015) - npm availability, new on() semantics.
