+++ browser.js
@@ -1,0 +1,3 @@
+import Pollymer from './src/Pollymer.js'
+
+window.Pollymer = Pollymer;
