+++ index.js
@@ -1,0 +1,2 @@
+import Pollymer from './src/Pollymer';
+export default Pollymer;
