+++ package.json
@@ -1,0 +1,35 @@
+{
+  "name": "pollymer",
+  "version": "1.1.3",
+  "description": "Pollymer Long-polling Ajax Library",
+  "main": "./commonjs.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "make",
+    "prepublish": "npm run build"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/fanout/pollymer.git"
+  },
+  "author": "fanout",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/fanout/pollymer/issues"
+  },
+  "homepage": "https://github.com/fanout/pollymer#readme",
+  "private": true,
+  "devDependencies": {
+    "babel-cli": "^6.2.4",
+    "babel-preset-es2015": "^6.1.18",
+    "jspm": "^0.16.15"
+  },
+  "jspm": {
+    "main": "./index.js",
+    "devDependencies": {
+      "babel": "npm:babel-core@^5.8.24",
+      "babel-runtime": "npm:babel-runtime@^5.8.24",
+      "core-js": "npm:core-js@^1.1.4"
+    }
+  }
+}
