+++ src/ErrorTypes.js
@@ -1,0 +1,4 @@
+export default {
+    "TransportError": 0,
+    "TimeoutError": 1
+};
