+++ src/Pollymer.js
@@ -1,0 +1,5 @@
+import Request from './Request';
+import ErrorTypes from './ErrorTypes';
+import TransportTypes from './TransportTypes';
+
+export default { Request, ErrorTypes, TransportTypes };
