+++ src/TransportTypes.js
@@ -1,0 +1,5 @@
+export default {
+    "Auto": 0,
+    "Xhr": 1,
+    "Jsonp": 2
+};
